Added a setNbLives to the Particle class so that any number of lives can be set
