+++ board/pathfinder.py
@@ -8,19 +8,19 @@ class UnreachableDestination(Exception):
     pass
 
 
-def get_shortest_path(board: Board, source_tile_id, dest_tile_id,  walkable_tile_function: function) -> (list, int):
+def get_shortest_path(source_tile_id, dest_tile_id, get_tile_by_id_func: function, get_tile_neighbours_func: function,
+                      walkable_tile_func: function) -> (list, int):
     """
     Uses the A* algorithm to find the shortest path between the source tile and
     the destination tile. This method is designed to be fast
 
     Args:
-        board: The board in which the path finding will be performed
         source_tile_id: The identifier of the tile from which the path finding is started
         dest_tile_id: The identifier of the tile to which the path finding ends
-        walkable_tile_function:
+        get_tile_by_id_func: function allowing to get the tile located at the given ID
+        get_tile_neighbours_func: function allowing to get the IDs of the given tile's neighbours.
+        walkable_tile_func:
             A function that take a tile in parameter and that returns True if the given tile can be walked on.
-            By default, only the "walkability" of the tile is tested, which means that a deadly tile is considered in
-            the path finding.
 
     Returns:
         The shortest path (list of tile_ids) to travel to reach the destination tile from the source tile.
@@ -33,72 +33,33 @@ def get_shortest_path(board: Board, source_tile_id, dest_tile_id,  walkable_tile
             (x1, y1) = a
             (x2, y2) = b
             return abs(x1 - x2) + abs(y1 - y2)
-        # TODO: Handle every other types of identifiers
 
-    frontier = PriorityQueue()
-    frontier.put(source_tile_id, 0)
-    came_from = {}
-    cost_so_far = {source_tile_id: 0}
-    came_from[source_tile_id] = None
-
-    while not frontier.empty():
-        current = frontier.get()
-
-        if current == dest_tile_id:
-            break
-        for next_tile_id in board.getTileById(current).neighbours:
-            new_cost = cost_so_far[current] + 1  # Cost just one more because all the cost in the graph == 1
-            next_tile = board.getTileById(next_tile_id)
-            if next_tile.walkable and (walkable_tile_function is None or walkable_tile_function(next_tile)) \
-                    and (next_tile_id not in cost_so_far or new_cost < cost_so_far[next_tile_id]):
-                cost_so_far[next_tile_id] = new_cost
-                priority = new_cost + heuristic(dest_tile_id, next_tile_id)
-                frontier.put(next_tile_id, priority)
-                came_from[next_tile_id] = current
+    came_from, _ = __find_path(dest_tile_id, source_tile_id, get_tile_by_id_func, get_tile_neighbours_func,
+                               walkable_tile_func, heuristic, -1)
     return reconstruct_path(came_from, source_tile_id, dest_tile_id)
 
 
-def get_shortest_paths(board: Board, source_tile_id, max_dist: int, walkable_tile_function: function) -> (dict, list):
+def get_shortest_paths(source_tile_id, max_dist: int, get_tile_by_id_func: function, get_tile_neighbours_func: function,
+                       walkable_tile_function: function) -> (dict, list):
     """
     Uses Dijkstra's algorithm to find all the paths available from the given source tile to any other tile
     within the given maximum distance.
 
     Args:
-        board: The board in which the algorithm will search the shortest paths
         source_tile_id: The identifier of the tile from which the path finding is started
         max_dist: The maximum distance allowed between the source and the destination tile (< 0 => no max distance)
+        get_tile_by_id_func: function allowing to get the tile located at the given ID
+        get_tile_neighbours_func: function allowing to get the IDs of the given tile's neighbours.
         walkable_tile_function:
             A function that take a tile in parameter and that returns True if the given tile can be walked on.
-            By default, only the "walkability" of the tile is tested, which means that a deadly tile is considered in
-            the path finding.
 
     Returns:
         A tuple containing:
             - A dict containing, for a key being a tile identifier, the previous tile
             - The cost to reach that destination tile
     """
-
-    frontier = PriorityQueue()
-    frontier.put(source_tile_id, 0)
-    came_from = {}
-    cost_so_far = {source_tile_id: 0}
-    came_from[source_tile_id] = None
-
-    while not frontier.empty():
-        current = frontier.get()
-
-        for next_tile_id in board.getTileById(current).neighbours:
-            new_cost = cost_so_far[current] + 1
-            next_tile = board.getTileById(next_tile_id)
-            if max_dist < 0 or new_cost <= max_dist:
-                if next_tile.walkable and (walkable_tile_function is None or walkable_tile_function(next_tile)) \
-                        and (next_tile_id not in cost_so_far or new_cost < cost_so_far[next_tile_id]):
-                    cost_so_far[next_tile_id] = new_cost
-                    priority = new_cost
-                    frontier.put(next_tile_id, priority)
-                    came_from[next_tile_id] = current
-
-    return came_from, cost_so_far
+    return __find_path(None, source_tile_id, get_tile_by_id_func, get_tile_neighbours_func, walkable_tile_function,
+                       None, max_dist)
 
 
 def reconstruct_path(came_from, source_tile_id, dest_tile_id):
@@ -113,3 +74,30 @@ def reconstruct_path(came_from, source_tile_id, dest_tile_id):
     except KeyError:
         raise UnreachableDestination("The tile " + str(dest_tile_id) +
                                      " is unreachable from the tile " + str(source_tile_id))
+
+
+def __find_path(dest_tile_id, source_tile_id, get_tile_by_id_func: function, get_tile_neighbours_func: function,
+                walkable_tile_function: function, heuristic, max_dist):
+    frontier = PriorityQueue()
+    frontier.put(source_tile_id, 0)
+    came_from = {}
+    cost_so_far = {source_tile_id: 0}
+    came_from[source_tile_id] = None
+    while not frontier.empty():
+        current = frontier.get()
+
+        if dest_tile_id is not None and current == dest_tile_id:
+            break
+        for next_tile_id in get_tile_neighbours_func(get_tile_by_id_func(current)):
+            new_cost = cost_so_far[current] + 1  # Cost just one more because all the cost in the graph == 1
+            next_tile = get_tile_by_id_func(next_tile_id)
+            if next_tile.walkable and (walkable_tile_function is None or walkable_tile_function(next_tile)) \
+                    and (next_tile_id not in cost_so_far or new_cost < cost_so_far[next_tile_id]):
+                cost_so_far[next_tile_id] = new_cost
+                priority = new_cost
+                if heuristic is not None:
+                    priority += heuristic(dest_tile_id, next_tile_id)
+                if max_dist < 0 or new_cost < max_dist:
+                    frontier.put(next_tile_id, priority)
+                came_from[next_tile_id] = current
+    return came_from, cost_so_far
